test(prune): split mega-test file into per-module suite
Comply with CONTRIBUTING.md 600-line cap. The original tests/plugins/test_prune_plugin.py was 1572 lines (2.6x over). Reorganized into a tests/plugins/prune/ package along module-under-test cohesion:

  _helpers.py              shared msg constructors + _menu_with_history (DRY: 125 lines that used to repeat at the top of the monolith now live exactly once)
